requirements.txt
yt-dlp==2023.12.30
customtkinter==5.2.1
pillow==10.2.0 